+++ __tests__/todo.spec.ts
@@ -172,80 +172,4 @@ testImplementations((api) => {
     console.log(result.errors);
     assertCompositionSuccess(result);
   });
-
-  test.todo("provides-on-union", () => {
-    const result = api.composeServices([
-      {
-        name: "a",
-        typeDefs: parse(/* GraphQL */ `
-          extend schema
-            @link(
-              url: "https://specs.apollo.dev/federation/v2.3"
-              import: ["@key", "@shareable", "@external"]
-            )
-
-          type Query {
-            media: [Media] @shareable
-          }
-
-          union Media = Book | Movie
-
-          type Book @key(fields: "id") {
-            id: ID!
-          }
-
-          type Movie @key(fields: "id") {
-            id: ID!
-          }
-        `),
-      },
-      {
-        name: "b",
-        typeDefs: parse(/* GraphQL */ `
-          extend schema
-            @link(
-              url: "https://specs.apollo.dev/federation/v2.3"
-              import: ["@key", "@shareable", "@provides", "@external"]
-            )
-
-          type Query {
-            media: [Media] @shareable @provides(fields: "... on Book { title }")
-          }
-
-          union Media = Book | Movie
-
-          type Book @key(fields: "id") {
-            id: ID!
-            title: String @external
-          }
-
-          type Movie @key(fields: "id") {
-            id: ID!
-          }
-        `),
-      },
-      {
-        name: "c",
-        typeDefs: parse(/* GraphQL */ `
-          extend schema
-            @link(
-              url: "https://specs.apollo.dev/federation/v2.3"
-              import: ["@key", "@shareable"]
-            )
-
-          type Book @key(fields: "id") {
-            id: ID!
-            title: String @shareable
-          }
-
-          type Movie @key(fields: "id") {
-            id: ID!
-            title: String @shareable
-          }
-        `),
-      },
-    ]);
-    console.log(result.errors);
-    assertCompositionSuccess(result);
-  });
 });
