+++ src/supergraph/validation/rules/satisfiablity/graph.ts
@@ -355,8 +355,12 @@ export class Graph {
         );
       }
 
+      const newTail = this.duplicateNode(edge.tail);
+      const newEdge = new Edge(edge.head, edge.move, newTail);
+      this.replaceEdgeAt(edge.head.index, edge.tail.index, newEdge, edgeIndex);
+
       queue.push({
-        head: edge.tail,
+        head: newTail,
         providedFields,
       });
     }
@@ -446,8 +450,12 @@ export class Graph {
         );
       }
 
+      const newTail = this.duplicateNode(edge.tail);
+      const newEdge = new Edge(edge.head, edge.move, newTail);
+      this.replaceEdgeAt(edge.head.index, edge.tail.index, newEdge, edgeIndex);
+
       queue.push({
-        head: edge.tail,
+        head: newTail,
         providedFields,
       });
     }
@@ -660,12 +668,16 @@ export class Graph {
 
         // find field edges that are provided and mark them as resolvable.
         // if they are not available, create them
+        const newTail = this.duplicateNode(edge.tail);
+        const newEdge = new Edge(edge.head, edge.move, newTail);
+        this.replaceEdgeAt(headIndex, edge.tail.index, newEdge, edgeIndex);
+
         const queue: {
           head: Node;
           providedFields: SelectionNode[];
         }[] = [
           {
-            head: edge.tail,
+            head: newTail,
             providedFields: edge.move.provides.selectionSet,
           },
         ];
